+++ src/main/java/com/coagronet/tipoBloque/dtos/TipoBloqueDTO.java
@@ -11,6 +11,6 @@ public class TipoBloqueDTO {
     private Integer id;
     private String nombre;
     private String descripcion;
-    private Integer estado;
+    private Long estado;
     private Long empresa;
 }
